Multiple issues with getting-started.md
Mention the black pins that should hold the ribbon cable in.

Add warning that "Arduino ESP32 Boards" is not the same as 'esp32'. I made this mistake, presumably others will too.

Latest v2 version of esp32 is v2.0.18, I assume this is OK.

I'd like to explain more about 'Choose an example' but I'm not sure what to say.

I've addressed some of the minor grammatical errors.